been working on seeding

+++ Cakefile
@@ -7,10 +7,13 @@ couchy = require './lib/couchy'
 noop = ->
 destroyAll = ->
   couchy('couchy-test-seed').destroy()
+  couchy('couchy-test-seed2').destroy()
   couchy('couchy-test-setup').destroy()
 
 logTask = (msg) ->
   console.log "\n#{msg}"
+
+option '-n', '--noclean', 'Do not destroy databases after test'
 
 task 'docs', 'Generate documentation', (options) ->
   logTask "Generating documentation"
@@ -23,7 +26,7 @@ task 'test', 'Run the test suites', (options) ->
     console.error err if err?
     console.error stderr
     console.log stdout
-    destroyAll()
+    destroyAll() unless options.noclean
 
 task 'cleanup', 'Deletes the database', (options) ->
   logTask "Deleting test database"
